+++ lib/glimmer/libui/control_proxy/triple_column.rb
@@ -1,0 +1,45 @@
+# Copyright (c) 2021 Andy Maleh
+#
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+#
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+# LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+module Glimmer
+  module LibUI
+    class ControlProxy
+      # A dual column is one that represents two values (e.g. image and text or checkbox and text)
+      # It is meant to be included in a column proxy class that already includes Column
+      module TripleColumn
+        def second_column_index
+          column_index + 1
+        end
+        
+        def third_column_index
+          column_index + 2
+        end
+        
+        def column_index
+          @column_index ||= @parent_proxy.send(:next_column_index).tap do
+            @parent_proxy.send(:next_column_index)
+            @parent_proxy.send(:next_column_index)
+          end
+        end
+      end
+    end
+  end
+end